feat: enable cache by default
